+++ CHANGELOG.md
@@ -176,6 +176,7 @@
     - **Dependencies:** Bump STACKIT SDK core module to `v0.26.0`
   - [v1.13.0](services/loadbalancer/CHANGELOG.md#v1130)
     - **Improvement:** Use new WaiterHelper for LoadBalancer waiters
+    - **Breaking Change:** `v2api/wait/DeleteLoadBalancerWaitHandler` now returns a `LoadBalancer` instead of `struct{}`
 - `logme`: 
   - [v0.27.3](services/logme/CHANGELOG.md#v0273) 
     - **Dependencies:** Bump STACKIT SDK core module from `v0.24.0` to `v0.24.1`